updated NumOfCalculations to uint64
Changed number of calculations to uint64 and had to add a second int variable for Blueprints as they do not play well with each other.
Blueprint now increments its own variable and updates NumOfCalculations on the CPP side during results collection.

This was done so we can do a higher number of calculations to get more detailed results on the CPP side for future tests.